+++ example/http-server-fast/http_server_fast.cpp
@@ -7,6 +7,8 @@
 
 #include "fields_alloc.hpp"
 
+#include "../common/mime_types.hpp"
+
 #include <beast/core.hpp>
 #include <beast/http.hpp>
 #include <beast/version.hpp>
@@ -23,37 +25,6 @@
 namespace ip = boost::asio::ip; // from <boost/asio.hpp>
 using tcp = boost::asio::ip::tcp; // from <boost/asio.hpp>
 namespace http = beast::http; // from <beast/http.hpp>
-
-// Return a reasonable mime type based on the extension of a file.
-//
-beast::string_view
-mime_type(boost::filesystem::path const& path)
-{
-    using beast::iequals;
-    auto const ext = path.extension().string();
-    if(iequals(ext, ".txt"))  return "text/plain";
-    if(iequals(ext, ".htm"))  return "text/html";
-    if(iequals(ext, ".html")) return "text/html";
-    if(iequals(ext, ".php"))  return "text/html";
-    if(iequals(ext, ".css"))  return "text/css";
-    if(iequals(ext, ".js"))   return "application/javascript";
-    if(iequals(ext, ".json")) return "application/json";
-    if(iequals(ext, ".xml"))  return "application/xml";
-    if(iequals(ext, ".swf"))  return "application/x-shockwave-flash";
-    if(iequals(ext, ".flv"))  return "video/x-flv";
-    if(iequals(ext, ".png"))  return "image/png";
-    if(iequals(ext, ".jpe"))  return "image/jpeg";
-    if(iequals(ext, ".jpeg")) return "image/jpeg";
-    if(iequals(ext, ".jpg"))  return "image/jpeg";
-    if(iequals(ext, ".gif"))  return "image/gif";
-    if(iequals(ext, ".bmp"))  return "image/bmp";
-    if(iequals(ext, ".ico"))  return "image/vnd.microsoft.icon";
-    if(iequals(ext, ".tiff")) return "image/tiff";
-    if(iequals(ext, ".tif"))  return "image/tiff";
-    if(iequals(ext, ".svg"))  return "image/svg+xml";
-    if(iequals(ext, ".svgz")) return "image/svg+xml";
-    return "application/text";
-}
 
 class http_worker
 {
@@ -231,7 +202,7 @@ private:
         // Fill out the reply to be sent to the client.
         response_.set(http::field::content_type, mime_type(target.to_string()));
         response_.body.clear();
-        for (char buf[512]; is.read(buf, sizeof(buf)).gcount() > 0;)
+        for (char buf[2048]; is.read(buf, sizeof(buf)).gcount() > 0;)
             response_.body.append(buf, static_cast<std::size_t>(is.gcount()));
         response_.prepare_payload();
     }