+++ include/beast/http/buffer_body.hpp
@@ -96,21 +96,14 @@ struct buffer_body
         value_type const& body_;
 
     public:
-        using is_deferred = std::false_type;
-
         using const_buffers_type =
             boost::asio::const_buffers_1;
 
         template<bool isRequest, class Fields>
         explicit
         reader(message<isRequest, buffer_body,
-                Fields> const& msg)
+                Fields> const& msg, error_code& ec)
             : body_(msg.body)
-        {
-        }
-
-        void
-        init(error_code& ec)
         {
             ec.assign(0, ec.category());
         }
@@ -145,12 +138,6 @@ struct buffer_body
                 ec.assign(0, ec.category());
             return boost::none;
         }
-
-        void
-        finish(error_code& ec)
-        {
-            ec.assign(0, ec.category());
-        }
     };
 #endif
 
@@ -165,13 +152,10 @@ struct buffer_body
     public:
         template<bool isRequest, class Fields>
         explicit
-        writer(message<isRequest, buffer_body, Fields>& m)
+        writer(message<isRequest, buffer_body, Fields>& m,
+                boost::optional<std::uint64_t> const&,
+                    error_code& ec)
             : body_(m.body)
-        {
-        }
-
-        void
-        init(boost::optional<std::uint64_t>, error_code& ec)
         {
             ec.assign(0, ec.category());
         }