+++ include/beast/http/string_view_body.hpp
@@ -48,21 +48,14 @@ struct string_view_body
         string_view body_;
 
     public:
-        using is_deferred = std::false_type;
-
         using const_buffers_type =
             boost::asio::const_buffers_1;
 
         template<bool isRequest, class Fields>
         explicit
-        reader(message<
-                isRequest, string_view_body, Fields> const& m)
+        reader(message<isRequest, string_view_body,
+                Fields> const& m, error_code& ec)
             : body_(m.body)
-        {
-        }
-
-        void
-        init(error_code& ec)
         {
             ec.assign(0, ec.category());
         }
@@ -72,12 +65,6 @@ struct string_view_body
         {
             ec.assign(0, ec.category());
             return {{{body_.data(), body_.size()}, false}};
-        }
-
-        void
-        finish(error_code& ec)
-        {
-            ec.assign(0, ec.category());
         }
     };
 #endif